test/abi: clean up test to fix builders
go.mod file was not tidy, made builders sad.

Updates #40724.

Reviewed-on: https://go-review.googlesource.com/c/go/+/308590
Trust: David Chase <drchase@google.com>
Run-TryBot: David Chase <drchase@google.com>
