--- test/abi/bad_select_crash.dir/go.mod
@@ -1,2 +1,0 @@
-module bad_select_crash.dir
-go 1.14
